+++ R/methods.R
@@ -491,20 +491,23 @@ plot.vsel <- function(
 #'   * `"mlpd"`: mean log predictive density, that is, `"elpd"` divided by the
 #'   number of observations.
 #'   * `"mse"`: mean squared error.
-#'   * `"rmse"`: root mean squared error. For the corresponding standard error,
+#'   * `"rmse"`: root mean squared error. For the corresponding standard error
+#'   (**TODO (bootstrap)**: and lower and upper confidence interval bounds?),
 #'   bootstrapping is used.
 #'   * `"acc"` (or its alias, `"pctcorr"`): classification accuracy
 #'   ([binomial()] family only).
 #'   * `"auc"`: area under the ROC curve ([binomial()] family only). For the
-#'   corresponding standard error, bootstrapping is used.
+#'   corresponding standard error and lower and upper confidence interval
+#'   bounds, bootstrapping is used.
 #' @param type One or more items from `"mean"`, `"se"`, `"lower"`, `"upper"`,
 #'   `"diff"`, and `"diff.se"` indicating which of these to compute for each
 #'   item from `stats` (mean, standard error, lower and upper confidence
 #'   interval bounds, mean difference to the corresponding statistic of the
 #'   reference model, and standard error of this difference, respectively). The
-#'   confidence interval bounds belong to normal-approximation confidence
-#'   intervals with (nominal) coverage `1 - alpha`. Items `"diff"` and
-#'   `"diff.se"` are only supported if `deltas` is `FALSE`.
+#'   confidence interval bounds belong to normal-approximation (or bootstrap;
+#'   see argument `stats`) confidence intervals with (nominal) coverage `1 -
+#'   alpha`. Items `"diff"` and `"diff.se"` are only supported if `deltas` is
+#'   `FALSE`.
 #' @param deltas If `TRUE`, the submodel statistics are estimated as differences
 #'   from the baseline model (see argument `baseline`) instead of estimating the
 #'   actual values of the statistics.