+++ tests/testthat/helpers/testers.R
@@ -302,12 +302,12 @@ refmodel_tester <- function(
 
   # dis
   if (refmod$family$family == "gaussian") {
-    expect_true(is.vector(refmod$dis, "double"), info = info_str)
-    expect_length(refmod$dis, nrefdraws_expected)
-    if (!is_datafit) {
-      expect_true(all(refmod$dis > 0), info = info_str)
-    } else {
+    if (is_datafit) {
       expect_identical(refmod$dis, 0, info = info_str)
+    } else {
+      expect_true(is.vector(refmod$dis, "double"), info = info_str)
+      expect_length(refmod$dis, nrefdraws_expected)
+      expect_true(all(refmod$dis > 0), info = info_str)
     }
   } else {
     expect_identical(refmod$dis, rep(NA, nrefdraws_expected), info = info_str)
@@ -730,7 +730,7 @@ submodl_tester <- function(
                        info = info_str)
 
       # coef()
-      coefs_crr <- coef(submodl_totest[[!!j]])
+      coefs_crr <- coef(submodl_totest[[j]])
       expect_type(coefs_crr, "list")
       expect_length(coefs_crr, length(nlvl_ran))
       for (zz in seq_len(length(nlvl_ran))) {
@@ -1305,14 +1305,14 @@ vsel_tester <- function(
       expect_true(all(!is.na(vs$summaries$sub[[!!j]]$lppd)), info = info_str)
     }
     if (with_cv) {
-      expect_type(vs$summaries$sub[[!!j]]$w, "double")
-      expect_length(vs$summaries$sub[[!!j]]$w, nobsv)
-      expect_true(all(!is.na(vs$summaries$sub[[!!j]]$w)), info = info_str)
+      expect_type(vs$summaries$sub[[!!j]]$wcv, "double")
+      expect_length(vs$summaries$sub[[!!j]]$wcv, nobsv)
+      expect_true(all(!is.na(vs$summaries$sub[[!!j]]$wcv)), info = info_str)
       if (nloo_expected == nobsv) {
-        expect_equal(vs$summaries$sub[[!!j]]$w, rep(1 / nobsv, nobsv),
+        expect_equal(vs$summaries$sub[[!!j]]$wcv, rep(1 / nobsv, nobsv),
                      info = info_str)
       } else {
-        expect_true(any(vs$summaries$sub[[!!j]]$w != rep(1 / nobsv, nobsv)),
+        expect_true(any(vs$summaries$sub[[!!j]]$wcv != rep(1 / nobsv, nobsv)),
                     info = info_str)
       }
     }
@@ -1641,8 +1641,7 @@ smmry_sel_tester <- function(
           )
         }
       } else {
-        expect_equal(smmry_sel[, diff_nm[stat_idx]], numeric(nrow(smmry_sel)),
-                     info = info_str)
+        expect_true(all(is.na(smmry_sel[, diff_nm[stat_idx]])), info = info_str)
       }
     }
   }
@@ -1653,9 +1652,12 @@ smmry_sel_tester <- function(
       lower_nm <- paste(stats_expected, "lower", sep = ".")
     }
     for (stat_idx in seq_along(stats_expected)) {
-      expect_true(all(smmry_sel[, stats_mean_name[stat_idx]] >=
-                        smmry_sel[, lower_nm[stat_idx]]),
-                  info = info_str)
+      if (!stats_expected[stat_idx] %in% c("rmse", "auc")) {
+        # RMSE and AUC are excluded here because of PR #347.
+        expect_true(all(smmry_sel[, stats_mean_name[stat_idx]] >=
+                          smmry_sel[, lower_nm[stat_idx]]),
+                    info = info_str)
+      }
     }
   }
   if ("upper" %in% type_expected) {
@@ -1665,9 +1667,12 @@ smmry_sel_tester <- function(
       upper_nm <- paste(stats_expected, "upper", sep = ".")
     }
     for (stat_idx in seq_along(stats_expected)) {
-      expect_true(all(smmry_sel[, stats_mean_name[stat_idx]] <=
-                        smmry_sel[, upper_nm[stat_idx]]),
-                  info = info_str)
+      if (!stats_expected[stat_idx] %in% c("rmse", "auc")) {
+        # RMSE and AUC are excluded here because of PR #347.
+        expect_true(all(smmry_sel[, stats_mean_name[stat_idx]] <=
+                          smmry_sel[, upper_nm[stat_idx]]),
+                    info = info_str)
+      }
     }
   }