+++ tests/testthat/test_refmodel.R
@@ -86,7 +86,8 @@ test_that("offsets specified via argument `offset` work", {
     tolerance = 1e-12,
     info = "rstanarm.glm.gauss.stdformul.with_wobs.with_offs"
   )
-  nms_compare <- c("mu", "dis", "y", "loglik", "wobs", "wsample", "offset")
+  nms_compare <- c("mu", "eta", "dis", "y", "loglik", "intercept", "wobs",
+                   "wsample", "offset")
   expect_equal(
     refmod_offs_arg[nms_compare],
     refmods$rstanarm.glm.gauss.stdformul.with_wobs.with_offs[nms_compare],