+++ NEWS.md
@@ -8,6 +8,7 @@ If you read this from a place other than <https://mc-stan.org/projpred/news/inde
 
 * Several improvements in the documentation.
 * For the RMSE as well as the AUC (see argument `stats` of `summary.vsel()`), the bootstrapping results are now also used for inferring the lower and upper confidence interval bounds. (GitHub: #318, #347; thanks to users @awd97 and @VisionResearchBlog)
+* For `datafit`s, offsets are not supported anymore. (GitHub: #186 (partly), #351)
 
 ## Bug fixes